Bump to stable dependencies and make the non-optional (#42)
In guava 28 the Futures API has been stabilized. As such, we should
expect that futures-extra will work across future guava versions. Thus,
there is less risk to keeping dependencies as non-optional (and thus
be backwards compatible with 4.1.1,
https://github.com/spotify/futures-extra/issues/38).

This change also bumps api-commons which includes a change in the JPMS
module name.